修改 order controller 和 service

+++ src/main/java/com/dessertoasis/demo/controller/order/OrderController.java
@@ -1,7 +1,5 @@
 package com.dessertoasis.demo.controller.order;
 
-import java.time.LocalDateTime;
-import java.util.ArrayList;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -13,20 +11,10 @@ import com.dessertoasis.demo.model.cart.CartListDTO;
 import com.dessertoasis.demo.model.cart.CourseCartDTO;
 import com.dessertoasis.demo.model.cart.ProductCartDTO;
 import com.dessertoasis.demo.model.cart.ReservationCartDTO;
-import com.dessertoasis.demo.model.course.Course;
 import com.dessertoasis.demo.model.member.Member;
-import com.dessertoasis.demo.model.order.CourseOrderItem;
 import com.dessertoasis.demo.model.order.Order;
-import com.dessertoasis.demo.model.order.ProdOrderItem;
-import com.dessertoasis.demo.model.order.Reservation;
-import com.dessertoasis.demo.model.product.Product;
-import com.dessertoasis.demo.service.MemberService;
-import com.dessertoasis.demo.service.ProductService;
 import com.dessertoasis.demo.service.cart.CartService;
-//import com.dessertoasis.demo.service.classroom.ReservationService;
-import com.dessertoasis.demo.service.course.CourseService;
 import com.dessertoasis.demo.service.order.OrderService;
-import com.dessertoasis.demo.service.order.ReservationService;
 
 import jakarta.servlet.http.HttpSession;
 
@@ -37,18 +25,6 @@ public class OrderController {
 	private OrderService orderService;
 	
 	@Autowired
-	private MemberService memberService;
-	
-	@Autowired
-	private ProductService productServie;
-	
-	@Autowired
-	private CourseService courseService;
-	
-	@Autowired
-	private ReservationService reservationService;
-	
-	@Autowired
 	private CartService cartService;
 	
 	// 新增訂單
@@ -56,57 +32,28 @@ public class OrderController {
 	public String insertOrder(@RequestBody CartListDTO cartList, HttpSession session) {		
 		Member member = (Member) session.getAttribute("loggedInMember");
 		if (member == null) return "沒有會員";
-		member = memberService.findByMemberId(member.getId());
-		
-		Order order = new Order();
-		order.setMember(member);
-		
+				
 		List<ProductCartDTO> productCartDTOs = cartList.getProductCartDTOs();
 		List<CourseCartDTO> courseCartDTOs = cartList.getCourseCartDTOs();
 		List<ReservationCartDTO> rsvCartDTOs = cartList.getReservationCartDTOs();
-		 
-		if (productCartDTOs != null) {
-			order.setProdOrderItems(new ArrayList<>());
-			for (ProductCartDTO cartDTO : productCartDTOs) {
-				Product product = productServie.findProductById(cartDTO.getProductId());
-				ProdOrderItem ordItem = new ProdOrderItem(cartDTO, product, order);
-				order.getProdOrderItems().add(ordItem);
-			}
+		
+		// 檢查教室是否已被預約
+		String rsvCheckResult = orderService.checkReservation(rsvCartDTOs);
+		if (rsvCheckResult != null) {
+			return rsvCheckResult;
 		}
 		
-		if (courseCartDTOs != null) {
-			order.setCourseOrderItems(new ArrayList<>());
-			for (CourseCartDTO cartDTO : courseCartDTOs) {
-				Course course = courseService.findById(cartDTO.getCourseId());
-				CourseOrderItem ordItem = new CourseOrderItem(cartDTO, course, order);
-				order.getCourseOrderItems().add(ordItem);
-			}
-		}
+		Order order = new Order();
 		
-		if (rsvCartDTOs != null) {
-			order.setReservations(new ArrayList<>());
-			for (ReservationCartDTO cartDTO : rsvCartDTOs) {
-				// 查此教室某天某時段是否已經預約
-				Reservation rsv = reservationService.getByRoomIdWithDateAndTime(
-						cartDTO.getClassroom().getId(), cartDTO.getReservationDate(), cartDTO.getReservationTime());
-				if (rsv != null) {
-					String room = rsv.getClassroom().getRoomName();
-					String date = rsv.getReservationDate().toString();
-					String time = reservationService.timeMap(rsv.getReservationTime());
-					return room + "-" + date + "-" + time + " 已被預約";
-				}
-				
-				rsv = new Reservation(cartDTO, order);
-				order.getReservations().add(rsv);
-			}
-		}
+		// 訂單設置 orderItem
+		order = orderService.placeProdOrderItem(order, productCartDTOs);
+		order = orderService.placeCourseOrderItem(order, courseCartDTOs);
+		order = orderService.placeReservation(order, rsvCartDTOs);
 		
-		LocalDateTime currentDateTime = LocalDateTime.now();
-		order.setOrdDate(currentDateTime);
-		order.setUpdateDate(currentDateTime);
-		order.setOrdStatus("訂單成立");
+		// 新增訂單
+		orderService.insert(order, member.getId());
 		
-		orderService.insert(order);
+		// 清掉購物車
 		cartService.deleteCarts(productCartDTOs, courseCartDTOs, rsvCartDTOs);
 		
 		return "訂單新增成功";
